FR 2: User only access through login
To ensure that the users can only login through the login page,
updated the profile page to check the location - if the location
is invalid (ie - just typed into the URL), then it will just
redirect to the homepage.

Additionally, set up the login to pass the user's credentials to lookup
for populating the dashboard.

Additionally, if they try to add something weird into the URL, I added
a "404 Page not Found" page which will re-direct all random requests to
my "NotFound" page.