FIX: Update .gitignore to track builds
Now that the entire project is being pushed and tracked by heroku,
we need to ensure new builds are created prior to pushes and tracked
by git. Therefore, every time the repo is pulled down, heroku recognizes
the new build and pushes to the server

The .gitignore removed the /server/build ignore from the line so
that new builds are tracked. Additionally, we're pushing in the
latest build with this.